Refactor PBS integration tests into individual test modules for each builder API endpoint
